src/requirements.txt
Flask
Mysql-connector